Final Changes
In the original program, when the program asks either the user wants to readme file or no then if we type any other words it will continue to the next program which is asking about the recipe we want to add. I changed the if statement after line case '1' become while loop. Next, I added else if function that if the user type no then the program will do nothing. Then, if the user type other than yes or no the program will go back to the question.

In the original program, when the database is empty and the user press 2 to search the recipe, the program will give us a segmentation fault. What I did was I changed a little for the condition inside if statement, else if statement, and the while loop in the else statement. Therefore, there will be no segmentation fault anymore and also if the user just type a blank space in the function number 1 then when the user wants to search the recipe, the program will say that the recipe was not found.

In the original program, when the database is empty a